backend/requirements.txt
django==4.2.3
tzdata
sentry-sdk
dj-database-url
django-ninja
six
django-ninja-extra
django-ninja-jwt
